+++ openquake/hazardlib/source/base.py
@@ -18,11 +18,13 @@ Module :mod:`openquake.hazardlib.source.base` defines a base class for
 seismic sources.
 """
 import abc
+import math
 import zlib
 import numpy
 from openquake.baselib import general
 from openquake.hazardlib import mfd
-from openquake.hazardlib.geo import Point
+from openquake.hazardlib.geo import Point, geodetic
+from openquake.hazardlib.geo.surface.planar import PlanarSurface
 from openquake.hazardlib.source.rupture import ParametricProbabilisticRupture
 
 EPS = .01  # used for src.nsites outside the maximum_distance
@@ -37,6 +39,93 @@ def get_code2cls():
         if hasattr(cls, 'code'):
             dic[cls.code] = cls
     return dic
+
+
+def _array_hc(usd, lsd, mag, dims, strike, dip, clon, clat, cdep):
+    # from the rupture center we can now compute the coordinates of the
+    # four coorners by moving along the diagonals of the plane. This seems
+    # to be better then moving along the perimeter, because in this case
+    # errors are accumulated that induce distorsions in the shape with
+    # consequent raise of exceptions when creating PlanarSurface objects
+    # theta is the angle between the diagonal of the surface projection
+    # and the line passing through the rupture center and parallel to the
+    # top and bottom edges. Theta is zero for vertical ruptures (because
+    # rup_proj_width is zero)
+    array = numpy.zeros((3, 4))
+    half_length, half_width, half_height = dims / 2.
+    rdip = math.radians(dip)
+
+    # precalculated azimuth values for horizontal-only and vertical-only
+    # moves from one point to another on the plane defined by strike
+    # and dip:
+    azimuth_right = strike
+    azimuth_down = (azimuth_right + 90) % 360
+    azimuth_left = (azimuth_down + 90) % 360
+    azimuth_up = (azimuth_left + 90) % 360
+
+    # half height of the vertical component of rupture width
+    # is the vertical distance between the rupture geometrical
+    # center and it's upper and lower borders:
+    # calculate how much shallower the upper border of the rupture
+    # is than the upper seismogenic depth:
+    vshift = usd - cdep + half_height
+    # if it is shallower (vshift > 0) than we need to move the rupture
+    # by that value vertically.
+    if vshift < 0:
+        # the top edge is below upper seismogenic depth. now we need
+        # to check that we do not cross the lower border.
+        vshift = lsd - cdep - half_height
+        if vshift > 0:
+            # the bottom edge of the rupture is above the lower seismo
+            # depth; that means that we don't need to move the rupture
+            # as it fits inside seismogenic layer.
+            vshift = 0
+        # if vshift < 0 than we need to move the rupture up.
+
+    # now we need to find the position of rupture's geometrical center.
+    # in any case the hypocenter point must lie on the surface, however
+    # the rupture center might be off (below or above) along the dip.
+    if vshift != 0:
+        # we need to move the rupture center to make the rupture fit
+        # inside the seismogenic layer.
+        hshift = abs(vshift / math.tan(rdip))
+        clon, clat = geodetic.point_at(
+            clon, clat, azimuth_up if vshift < 0 else azimuth_down,
+            hshift)
+        cdep += vshift
+    theta = math.degrees(math.atan(half_width / half_length))
+    hor_dist = math.sqrt(half_length ** 2 + half_width ** 2)
+    azimuths = numpy.array([(strike + 180 + theta) % 360,
+                            (strike - theta) % 360,
+                            (strike + 180 - theta) % 360,
+                            (strike + theta) % 360])
+    array[:2] = geodetic.point_at(clon, clat, azimuths, hor_dist)
+    array[2, 0:2] = cdep - half_height
+    array[2, 2:4] = cdep + half_height
+    return array, numpy.array([clon, clat, cdep])
+
+
+def _get_surfaces(inp, hypo, shift_hypo=False):
+    """
+    :returns: a list of rupture surfaces
+    :param inp:
+        Surface input parameters
+    :param hypo:
+        Hypocenter
+    :param shift_hypo:
+        If true, change .hc to the shifted hypocenter
+    :return:
+        PlanarSurface instances with attribute .hc
+    """
+    out = []
+    for rec in inp:
+        array, hc = _array_hc(rec.usd, rec.lsd, rec.mag, rec.dims,
+                              rec.strike, rec.dip, hypo.x, hypo.y, hypo.z)
+        surface = PlanarSurface.from_array(  # shape (3, 4)
+            array, rec.strike, rec.dip)
+        surface.hc = Point(*hc) if shift_hypo else hypo
+        out.append(surface)
+    return out
 
 
 class BaseSeismicSource(metaclass=abc.ABCMeta):
@@ -189,7 +278,7 @@ class BaseSeismicSource(metaclass=abc.ABCMeta):
                 hc = Point(latitude=src.location.latitude,
                            longitude=src.location.longitude,
                            depth=hc_depth)
-                surface = src._get_rupture_surface(mag, np, hc)
+                [surface] = _get_surfaces(src.get_input([mag], np), hc)
                 rup = ParametricProbabilisticRupture(
                     mag, np.rake, src.tectonic_region_type, hc,
                     surface, rate, tom)
